Align inter-pod affinity closer to what Kubernetes does today
Kubernetes today makes an exception in the case of an inter-pod affinity
requirement R for a pod p under the following condition: when there are no
nodes that satisfy R, and p is affine to itself according to R, then p can
be placed on any node. This departs from an earlier version of the DCM
scheduler where such pods were never placed.

Signed-off-by: Lalith Suresh <lsuresh@vmware.com>