k8s-scheduler: Allow controllable__node_name to be NULL_NODE
Until now, the variable column used for identifying node placements
required assignments to valid nodes. This meant that if a single pod could
not be placed because of constraints being unmet, the model would fail
to place the entire batch of pods and throw an UNSAT.

With this commit, we allow pods to be assigned to "NULL_NODE", which
simply indicates an unassigned pod. Doing so also opens up a clean
path towards accounting for pre-emption and eviction/descheduling.

Signed-off-by: Lalith Suresh <lsuresh@vmware.com>