Adds a test for the KubernetesStateSync class
We've been previously exercising this code path through integration
tests, but they seem to be a pain to run on CirceCI. A simpler
alternative is to use the Kubernetes mock server that comes bundled
with our Kubernetes client (fabric8io). For now, we use the
expectation-based interface, which is more verbose than we'd like.
Once they add support for informers in CRUD-mode, we can simplify
KubernetesStateSyncTest.

Signed-off-by: Lalith Suresh <lsuresh@vmware.com>