Remove Nullability within visitor code
Dealing with new visitors is unnecessarily burdened by nulls
propagating through the various methods. This was originally
the case because the generic `MonoidVisitor<T, C>` base class
attempted to use null as defaults for `T` and `C`. This is not
necessarily the case.

Scanning uses of this class throughout the code base, we
find that many sub-classes are either simple visitors
(for example, of type `MonoidVisitor<Void, Void>`). Heavy
usage of `Void` as one of the generic parameters is often
a clear signal that that parameter is not used.

This patch introduces a `VoidType` parameter that we use
instead of the `Void` type because the latter doesn't allow
us to instantiate an instance that we could use as a default.
We then  introduce a `SimpleVisitor` class that is
of type `MonoidVisitor<VoidType, VoidType>` that is sufficient
for many of our simpler visitors. The `ComprehensionRewriter`
type also happened to benefit from a move away from `Void`
as its generic parameter.

Signed-off-by: Lalith Suresh <lsuresh@vmware.com>