Simplify translation from Presto SQL AST to IR AST
The translation was previously done in two passes, one to get
an expression stack and another to construct an AST by processing
the stack. This is unnecessary and the IR AST can instead be produced
with a recursive pass over the SQL AST. This both simplifies the code and
eliminates some spurious duplication as well.

Signed-off-by: Lalith Suresh <lsuresh@vmware.com>