Adds support for check clauses to specify hard constraints
This patch supports specifying check clauses to specify hard constraints.
A `check (expr)` clause appears after a query to assert that `expr` holds
true for every record returned by the query.

Despite the syntax change, we still generate the same IR. But this patch
opens up room to significantly simplify both the IR and the backends.

Signed-off-by: Lalith Suresh <lsuresh@vmware.com>