Add check expressions as a type of qualifier.
This is based on lessons learnt from another prototype where we
treated check expressions as a "head". Treating check expressions
as a head type breaks several re-writing passes that simply scan/modify
the qualifiers. An easier increment to shoot for is to explicitly
introduce a `CheckExpression` type which is a sub-type of Qualifier.

This means that several of our rewriting/analysis passes (in particular,
the ones that divide qualifiers into non-var and var types, which
many sections of the code depend on), work with some minor edits.

Signed-off-by: Lalith Suresh <lsuresh@vmware.com>