Add check expressions to IR as an explicit subtype of Qualifier (#55)
* Add check expressions as a type of qualifier.

This is based on lessons learnt from another prototype where we
treated check expressions as a "head". Treating check expressions
as a head type breaks several re-writing passes that simply scan/modify
the qualifiers. An easier increment to shoot for is to explicitly
introduce a `CheckExpression` type which is a sub-type of Qualifier.

This means that several of our rewriting/analysis passes (in particular,
the ones that divide qualifiers into non-var and var types, which
many sections of the code depend on), work with some minor edits.

Signed-off-by: Lalith Suresh <lsuresh@vmware.com>

* Expand IR GroupBy clauses to also support constant expressions

This allows us to (correctly) insert a `group by 1` whenever
aggregate functions are involved in a query, but there are no
explicit group bys specified.

Signed-off-by: Lalith Suresh <lsuresh@vmware.com>

* Add CheckQualifier file

Signed-off-by: Lalith Suresh <lsuresh@vmware.com>