Use hash-based joins when feasible
The compiler has enough structural information to construct and use
hash-based indexes to speed up joins in the generated code, rather
than always use nested for loops. This patch introduces a first cut
and doing so. It does not yet consider any dynamic information
in picking which relations to scan (e.g, pick the smallest
relation), and it only considers equality-based join predicates
which involve the "from" relation.

Signed-off-by: Lalith Suresh <lsuresh@vmware.com>