Prepare to migrate to more recent JDKs (#63)
Bump up dependencies and prepare for more recent JDKs.

The build and tests pass with JDK 14. This commit also organizes all dependency versions into a single gradle.properties file.

For JDK 15, Gradle has not yet migrated to Jacoco 0.8.6 (which has experimental support for JDK 15 class files). This means the code coverage stage of the build fails, but the DCM build and all the tests run fine.

* Bump up checkstyle and jacoco versions to work with JDK 15. Also patch new errors reported by checkstyle.
* Use a gradle properties file to specify dependency and plugin versions
* gradleGitProperties variable -> gitPropertiesVersion