ddlog-jooq: add a fetcher API to allow callers to control input data to models (Fixes #57)
Until now, the only way to supply input data to a model is via the
`model.updateData()` call. Internally, this uses the currently
available database context to iterate over every input table/view
and invoke `dbCtx.selectFrom(table).fetch()`.

While convenient, it is not necessarily the most flexible API. First,
it goes over the JOOQ/JDBC boundary on every call, leading to unnecessary
latencies. Second, it does not allow users to dynamically subset the data
to use for a given input table without modifying the schema. Third,
it does not allow the user to exercise control over how the data is
fetched for a given table (e.g., use prepared statements, use an in-memory
cache of the data without going through JOOQ, or even retrieve a Result<?>
entry using a custom data source like ddlog).

This patch adds an `updateData(Function<Table, Result<? extends Record<?>>)`
API. `updateData()` works as before and delegates to use a default fetcher
that simply executes `dbCtx.selectFrom(table).fetch()`. An example use of
the new API might be:

```
   model.updateData(table -> {
        if (table.getName().equalsIgnoreCase("SOME_CACHED_TABLE")) {
          return resultCache;
        } else {
          return dbCtx.selectFrom(table).fetch();
        }
   });
```

In doing so, this also fixes #57: we should now be able to use ddlog's
`DDlogJooqProvider` to directly return a materialized view and short-cut
all of JOOQ's deserialization logic on `select * from T` calls.

Signed-off-by: Lalith Suresh <lsuresh@vmware.com>