[#81] dcm: rework objective function syntax
While hard constraints had migrated to the check syntax (<relation> check <expr>),
we were still using the "objective_" prefix in the view name to signal that a view
represents an objective function. This changeset aligns the two types of constraints.

Now an objective function is represented using the following syntax:

 <scalar expression | column expression> maximize;

Fixes #81

Signed-off-by: Lalith Suresh <lsuresh@vmware.com>