ortools-backend: get assumptions to work end-to-end
Given https://github.com/google/or-tools/issues/2563,
we need to make sure there is only a single search worker
and there are no objectives configured when using the
assumptions capability.

For now, we configure assumptions anyway, knowing that
or-tools will ignore them. If the problem is indeed
UNSAT, we then modify the model to set only one search
worker and remove the objective function, before trying
again.

This seems to work end-to-end. It is not clear if this
has performance implications.

Signed-off-by: Lalith Suresh <lsuresh@vmware.com>