[#96] dcm: Use consistent syntax for hard constraints and objectives (#100)
With #82 being fixed, there is little reason to have a different query structure for hard and soft constraints. We can instead structure both constraints as a view that produces a set of records, followed by a CHECK or MAXIMIZE clause, followed by an expression. This would allow a user to easily switch between hard/soft constraints if they want.

Signed-off-by: Lalith Suresh <lsuresh@vmware.com>