dcm: simplify Model and ISolverBackend APIs (#114)
* dcm: remove updateData()
* dcm: remove all instances of updateData()
* dcm: remove DSLContext parameter from runSolver()
* dcm: remove Map<String, IRTable> parameter from ISolverBackend.runSolver()

`model.updateData()` was always called just before invoking `model.solve()`, a clear sign that they shouldn't be two separate calls to begin with. This patch folds the equivalent functionality into `model.solve()`, and also adds overloaded versions to support a custom fetcher.

In doing so, it became clear that the internal APIs to interact with the solvers did not need an explicit `ISolverBackend.generateDataCode()` phase either. The patch further simplifies the interaction with the solvers, making it such that once model generation is done, the only argument required to invoke the solver is a set of tables.

These changes combine to make the Model instance and the solver backend completely stateless.

Signed-off-by: Lalith Suresh <lsuresh@vmware.com>