At the beginning of SchedulerTest, compile the DDlog program w Scoped Views once, so that we never have to compile it again with each of the CartesianProductTests.
We also temporarily change the Scheduler constructor to add AutoScoped views to DDlog, even if the Scheduler isn't running in AutoScoped mode. This is so that we can use the same DDlog program (with scoping) for non-scoped tests as well.

Need to fix this eventually, as for real tests/workloads, we might want to turn off Autoscoping.

Signed-off-by: Amy Tai <amy.tai.2009@gmail.com>